requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.7.1=py37_0
astor=0.7.1=py_0
atomicwrites=1.3.0=py_0
attrs=19.1.0=py_0
binutils_impl_linux-64=2.31.1=h6176602_1
binutils_linux-64=2.31.1=h6176602_7
bzip2=1.0.8=h516909a_0
c-ares=1.15.0=h516909a_1001
ca-certificates=2019.6.16=hecc5488_0
certifi=2019.6.16=py37_1
cloudpickle=1.2.1=py_0
cycler=0.10.0=py_1
cytoolz=0.10.0=py37h516909a_0
dask-core=2.1.0=py_0
decorator=4.4.0=py_0
freetype=2.10.0=he983fc9_0
gast=0.2.2=py_0
gcc_impl_linux-64=7.3.0=habb00fd_1
gcc_linux-64=7.3.0=h553295d_7
grpcio=1.16.1=py37hf8bcb03_1
gxx_impl_linux-64=7.3.0=hdf63c60_1
gxx_linux-64=7.3.0=h553295d_7
h5py=2.9.0=nompi_py37hf008753_1102
hdf5=1.10.4=nompi_h3c11f04_1106
icu=64.2=he1b5a44_0
imageio=2.5.0=py37_0
importlib_metadata=0.18=py37_0
jpeg=9c=h14c3975_1001
keras=2.2.4=py37_1
keras-applications=1.0.7=py_1
keras-preprocessing=1.0.9=py_1
kiwisolver=1.1.0=py37hc9558a2_0
libblas=3.8.0=10_openblas
libcblas=3.8.0=10_openblas
libffi=3.2.1=he1b5a44_1006
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libgpuarray=0.7.6=h14c3975_1003
liblapack=3.8.0=10_openblas
libopenblas=0.3.6=h6e990d7_5
libpng=1.6.37=hed695b0_0
libprotobuf=3.9.0=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h57b8799_1003
lz4-c=1.8.3=he1b5a44_1001
mako=1.0.10=py_0
markdown=3.1.1=py_0
markupsafe=1.1.1=py37h14c3975_0
matplotlib-base=3.1.1=py37he7580a8_1
mock=3.0.5=py37_0
more-itertools=7.1.0=py_0
ncurses=6.1=hf484d3e_1002
networkx=2.3=py_0
numpy=1.16.4=py37h95a1406_0
olefile=0.46=py_0
openblas=0.3.6=h6e990d7_5
openssl=1.1.1c=h516909a_0
packaging=19.0=py_0
pandas=0.25.0=py37hb3f55d8_0
pillow=6.1.0=py37he7afcd5_0
pip=19.2.1=py37_0
pluggy=0.12.0=py_0
protobuf=3.9.0=py37he1b5a44_0
py=1.8.0=py_0
pygpu=0.7.6=py37h3010b51_1000
pyparsing=2.4.1.1=py_0
pytest=5.0.1=py37_0
python=3.7.3=h33d41f4_1
python-dateutil=2.8.0=py_0
pytz=2019.1=py_0
pywavelets=1.0.3=py37hd352d35_1
pyyaml=5.1.1=py37h516909a_0
readline=8.0=hf8c457e_0
scikit-image=0.15.0=py37hb3f55d8_2
scipy=1.3.0=py37h921218d_0
setuptools=41.0.1=py37_0
six=1.12.0=py37_1000
sqlite=3.29.0=hcee41ef_0
tensorboard=1.13.1=py37_0
tensorflow=1.13.1=py37_0
tensorflow-estimator=1.13.0=py_0
termcolor=1.1.0=py_2
theano=1.0.4=py37hf484d3e_1000
tk=8.6.9=hed695b0_1002
toolz=0.10.0=py_0
tornado=6.0.3=py37h516909a_0
wcwidth=0.1.7=py_1
werkzeug=0.15.5=py_0
wheel=0.33.4=py37_0
xz=5.2.4=h14c3975_1001
yaml=0.1.7=h14c3975_1001
zipp=0.5.1=py_0
zlib=1.2.11=h516909a_1005
zstd=1.4.0=h3b9ef0a_0
